rename models/views to django conventions
