+++ back/core/gym/models.py
@@ -31,3 +31,6 @@ class Gym(models.Model):
 
     def __str__(self):
         return self.name
+
+    class Meta:
+        db_table = 'core_gym'
