+++ back/core/gym/urls.py
@@ -1,5 +1,5 @@
 from rest_framework import routers
-from core.gym.view import GymViewSet
+from core.gym.views import GymViewSet
 
 gym_router = routers.DefaultRouter()
 gym_router.register('gyms', GymViewSet)
