+++ back/core/gym_details/urls.py
@@ -1,5 +1,5 @@
 from rest_framework import routers
-from core.gym_details.view import GymDetailsViewSet
+from core.gym_details.views import GymDetailsViewSet
 
 gym_details_router = routers.DefaultRouter()
 gym_details_router.register('gym_details', GymDetailsViewSet, basename='gym_details')
