+++ back/core/gym_leaderboard/models.py
@@ -1,16 +1,21 @@
 from django.contrib.postgres.fields import ArrayField, JSONField
 from django.db import models
-from core.gym.model import Gym
+from core.gym.models import Gym
 
 
-class GymDetails(models.Model):
+class GymLeaderboard(models.Model):
+    leaderboard_api_id = models.IntegerField(blank=True, null=True)
     gym = models.OneToOneField(
         Gym,
         on_delete=models.PROTECT,
         primary_key=True,
     )
-    gym_details_api_id = models.TextField()
-    data = JSONField()
+    data = ArrayField(
+        JSONField()
+    )
 
     def __str__(self):
         return ""
+
+    class Meta:
+        db_table = 'core_gymleaderboard'
