+++ back/core/gym_leaderboard/urls.py
@@ -1,5 +1,5 @@
 from rest_framework import routers
-from core.gym_leaderboard.view import GymLeaderboardViewSet
+from core.gym_leaderboard.views import GymLeaderboardViewSet
 
 gym_leaderboard_router = routers.DefaultRouter()
 gym_leaderboard_router.register('gym_leaderboard', GymLeaderboardViewSet, basename='gym_leaderboard')
