+++ back/core/tests.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
-from core.gym.model import Gym
+from core.gym.models import Gym
 
 
 class GymViewTestCases(APITestCase):
